Use /usr/bin/env bash
Some systems don't have bash in /bin/bash. Notably the BSDs don't,
and some systems don't necessarily have an up-to-date version in /bin.

Watchman still doesn't work on FreeBSD, but this allows the build to
succeed with a more-or-less stock FreeBSD 11 installation, with the
following packages installed:

bash
openjdk8
python2